worker/pyproject.toml
[tool.poetry]
name = "ecommerce-worker"
version = "0.1.0"
description = ""
authors = ["DavinHHiu <honghieu2208@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1"
djangorestframework = "^3.15.2"
pillow = "^10.4.0"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
